+++ RELEASE_NOTES.md
@@ -3,5 +3,5 @@ v0.67.4
 - Android and iOS: ChatConnectionListener has a new optional callback for server alerts.
   (Already added for Node in v0.67.2.)
 
-- Node: Net.preconnectChat will start the connection process for an authenticated chat connection
+- Net.preconnectChat will start the connection process for an authenticated chat connection
   without needing a username and password ready.